+++ src/Hero/hero.module.scss
@@ -161,6 +161,7 @@ position: relative;
 }
 .buttonGap{
   display: flex;
+  align-items: center;
   gap: 1vw;
 }
 .content1Back {
